Add files via upload
Made many edits to most classes such as comments, fixed the issue of food not disappearing, also integrated the ghost class that Medhanie created. Still have an issue with location though as the food that gets consumed is the circles 90 to degrees to the left. Also there are issues with the text being added to the root.

Another major issue is that we are unable to move the characters more than 1 move in all directions.